Graphics.Rendering.Pango.Types

Description

This file reflects the Gtk+ object hierarchy in terms of Haskell classes.

Note: the mk... functions were originally meant to simply be an alias for the constructor. However, in order to communicate the destructor of an object to objectNew, the mk... functions are now a tuple containing Haskell constructor and the destructor function pointer. This hack avoids changing all modules that simply pass mk... to objectNew.

Documentation

module System.Glib.GObject

newtype PangoContext

Constructors

PangoContext (ForeignPtr PangoContext)

class GObjectClass o => PangoContextClass o

toPangoContext :: PangoContextClass o => o -> PangoContext

mkPangoContext :: (ForeignPtr PangoContext -> PangoContext, FinalizerPtr a)

unPangoContext :: PangoContext -> ForeignPtr PangoContext

castToPangoContext :: GObjectClass obj => obj -> PangoContext

gTypePangoContext :: GType

newtype PangoLayoutRaw

Constructors

PangoLayoutRaw (ForeignPtr PangoLayoutRaw)

class GObjectClass o => PangoLayoutRawClass o

toPangoLayoutRaw :: PangoLayoutRawClass o => o -> PangoLayoutRaw

mkPangoLayoutRaw :: (ForeignPtr PangoLayoutRaw -> PangoLayoutRaw, FinalizerPtr a)

unPangoLayoutRaw :: PangoLayoutRaw -> ForeignPtr PangoLayoutRaw

castToPangoLayoutRaw :: GObjectClass obj => obj -> PangoLayoutRaw

gTypePangoLayoutRaw :: GType

newtype Font

Constructors

Font (ForeignPtr Font)

class GObjectClass o => FontClass o

toFont :: FontClass o => o -> Font

mkFont :: (ForeignPtr Font -> Font, FinalizerPtr a)

unFont :: Font -> ForeignPtr Font

castToFont :: GObjectClass obj => obj -> Font

gTypeFont :: GType

newtype FontFamily

Constructors

FontFamily (ForeignPtr FontFamily)

class GObjectClass o => FontFamilyClass o

toFontFamily :: FontFamilyClass o => o -> FontFamily

mkFontFamily :: (ForeignPtr FontFamily -> FontFamily, FinalizerPtr a)

unFontFamily :: FontFamily -> ForeignPtr FontFamily

castToFontFamily :: GObjectClass obj => obj -> FontFamily

gTypeFontFamily :: GType

newtype FontFace

Constructors

FontFace (ForeignPtr FontFace)

class GObjectClass o => FontFaceClass o

toFontFace :: FontFaceClass o => o -> FontFace

mkFontFace :: (ForeignPtr FontFace -> FontFace, FinalizerPtr a)

unFontFace :: FontFace -> ForeignPtr FontFace

castToFontFace :: GObjectClass obj => obj -> FontFace

gTypeFontFace :: GType

newtype FontMap

Constructors

FontMap (ForeignPtr FontMap)

class GObjectClass o => FontMapClass o

toFontMap :: FontMapClass o => o -> FontMap

mkFontMap :: (ForeignPtr FontMap -> FontMap, FinalizerPtr a)

unFontMap :: FontMap -> ForeignPtr FontMap

castToFontMap :: GObjectClass obj => obj -> FontMap

gTypeFontMap :: GType

newtype FontSet

Constructors

FontSet (ForeignPtr FontSet)

class GObjectClass o => FontSetClass o

toFontSet :: FontSetClass o => o -> FontSet

mkFontSet :: (ForeignPtr FontSet -> FontSet, FinalizerPtr a)

unFontSet :: FontSet -> ForeignPtr FontSet

castToFontSet :: GObjectClass obj => obj -> FontSet

gTypeFontSet :: GType